first pass point 19

+++ utils.h
@@ -15,7 +15,7 @@ FILE *outputFileInit(FILE *outP, char *outPutFileName, char *inputName);
 void delay(int time); /* not necessary */
 
 void addSymbol(char *name, attribute *attribs, symbol *tempNode,
-               symbol *head, symbol *currNode, int *IC, char *line);
+               symbol *head, symbol *currNode, long *IC, char *line);
 
 char *iterator(char *line, FILE *inp, int *errors);
